requirement.txt
streamlit==1.33.0
pandas==2.2.1
numpy==1.24.4
matplotlib==3.7.1
mysql-connector-python==8.3.0
scikit-learn==1.3.2
pytesseract==0.3.10
openpyxl==3.1.2
transformers==4.40.0
torch==2.2.2
nltk==3.8.1
sastrawi==1.0.1
SpeechRecognition==3.10.1
pydub==0.25.1
opencv-python==4.9.0.80
python-docx==1.1.0
